rename to scenarios path
