+++ src/lib/components/info-graphics/scenarios-explorer/homepage/Preview.svelte
@@ -249,7 +249,7 @@
 			</a> -->
 
 			<a
-				href="/scenes"
+				href="/scenarios"
 				class="mt-12 md:mt-0 block text-center rounded-xl font-space border border-black border-solid p-6 transition-all text-white bg-black hover:bg-dark-grey hover:no-underline"
 			>
 				View scenario explorer
