--- src/routes/scenarios/PublicBetaTag.svelte
@@ -1,12 +1,0 @@
-<div class="bg-light-warm-grey py-6 md:py-4">
-	<div class="container max-w-none lg:container md:flex items-center gap-3 md:gap-6">
-		<span
-			class="inline-block whitespace-nowrap uppercase font-space font-medium tracking-widest text-xs bg-warm-grey py-2 px-4 rounded-full mb-2 md:mb-0"
-		>
-			Public Beta
-		</span>
-		<div class="text-mid-grey text-sm font-light">
-			This feature is in development. A more detailed Scenarios explorer will be available soon.
-		</div>
-	</div>
-</div>
