+++ src/routes/scenarios/components/ArticlesSection.svelte
