+++ src/routes/scenarios/components/DetailedRegion.svelte
