+++ src/routes/scenarios/components/DetailedScenario.svelte
