+++ src/routes/scenarios/components/DetailedTechnology.svelte
