+++ src/routes/scenarios/components/DownloadButton.svelte
