+++ src/routes/scenarios/components/MiniCharts.svelte
