+++ src/routes/scenarios/components/PathwaySelection.svelte
