+++ src/routes/scenarios/components/ScenarioButton.svelte
