+++ src/routes/scenarios/components/ScenarioChart.svelte
