+++ src/routes/scenarios/components/ScenarioDescription.svelte
