+++ src/routes/scenarios/components/ScenarioSelection.svelte
