+++ src/routes/scenarios/components/TableRegion.svelte
