+++ src/routes/scenarios/page-data-options/chart-ticks.js
