+++ src/routes/scenarios/page-data-options/exclude-battery-and-loads.js
