+++ src/routes/scenarios/page-data-options/view-sections.js
