+++ src/routes/scenarios/stores/data-viz.js
